feat: 알림 채널 생성 로직 -> Applciation 파일로 이동(메세지 수신할 때마다 채널 생성 방지)
